+++ xcube/cli/_gen2/open.py
@@ -23,9 +23,11 @@ from typing import Sequence
 
 from xcube.cli._gen2.genconfig import CubeConfig
 from xcube.cli._gen2.genconfig import InputConfig
+from xcube.core.normalize import normalize_dataset
 from xcube.core.store import DataStoreError
 from xcube.core.store import DataStorePool
 from xcube.core.store import TYPE_SPECIFIER_CUBE
+from xcube.core.store import TYPE_SPECIFIER_DATASET
 from xcube.core.store import get_data_store_instance
 from xcube.core.store import new_data_opener
 from xcube.util.progress import observe_progress
@@ -41,16 +43,32 @@ def open_cubes(input_configs: Sequence[InputConfig],
             opener_id = input_config.opener_id
             store_params = input_config.store_params or {}
             open_params = input_config.open_params or {}
+            normalisation_required = False
             if input_config.store_id:
                 store_instance = get_data_store_instance(input_config.store_id,
                                                          store_params=store_params,
                                                          store_pool=store_pool)
                 store = store_instance.store
                 if opener_id is None:
-                    opener_ids = store.get_data_opener_ids(data_id=input_config.data_id,
-                                                           type_specifier=TYPE_SPECIFIER_CUBE)
+                    opener_ids = None
+                    type_specifiers = store.get_type_specifiers_for_data(input_config.data_id)
+                    for type_specifier in type_specifiers:
+                        if TYPE_SPECIFIER_CUBE.is_satisfied_by(type_specifier):
+                            opener_ids = \
+                                store.get_data_opener_ids(data_id=input_config.data_id,
+                                                          type_specifier=type_specifier)
+                            break
                     if not opener_ids:
-                        raise DataStoreError(f'Data store "{input_config.store_id}" does not support data cubes')
+                        for type_specifier in type_specifiers:
+                            if TYPE_SPECIFIER_DATASET.is_satisfied_by(type_specifier):
+                                opener_ids = \
+                                    store.get_data_opener_ids(data_id=input_config.data_id,
+                                                              type_specifier=type_specifier)
+                                normalisation_required = True
+                                break
+                    if not opener_ids:
+                        raise DataStoreError(f'Data store "{input_config.store_id}" '
+                                             f'does not support datasets')
                     opener_id = opener_ids[0]
                 opener = store
                 open_params.update(opener_id=opener_id, **open_params)
@@ -60,6 +78,8 @@ def open_cubes(input_configs: Sequence[InputConfig],
             open_params_schema = opener.get_open_data_params_schema(input_config.data_id)
             cube_params = {k: v for k, v in all_cube_params.items() if k in open_params_schema.properties}
             cube = opener.open_data(input_config.data_id, **open_params, **cube_params)
+            if normalisation_required:
+                cube = normalize_dataset(cube)
             cubes.append(cube)
             progress.worked(1)